% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_dist_plot.R
\name{make_comb_dist_plot}
\alias{make_comb_dist_plot}
\title{Combine distribution plots}
\usage{
make_comb_dist_plot(risk_dist, cardioresp_dist, airway_dist, ncmscs_total_dist)
}
\description{
Combine distribution plots
}
